resiliency wip

+++ app/Jobs/Backup/DeleteJob.php
@@ -4,7 +4,9 @@ namespace App\Jobs\Backup;
 
 use App\DTOs\SocketEventDTO;
 use App\Enums\BackupFileStatus;
+use App\Enums\BackupStatus;
 use App\Events\SocketEvent;
+use App\Http\Resources\BackupResource;
 use App\Models\Backup;
 use App\Models\ServerLog;
 use App\Traits\UniqueQueue;
@@ -25,11 +27,23 @@ class DeleteJob implements ShouldQueue
             $projectId = $this->backup->server->project_id;
             $backupId = $this->backup->id;
 
-            $files = $this->backup->files;
-            foreach ($files as $file) {
+            foreach ($this->backup->files as $file) {
                 $file->status = BackupFileStatus::DELETING;
                 $file->save();
                 $file->deleteFile();
+            }
+
+            if ($this->backup->files()->exists()) {
+                $this->backup->status = BackupStatus::DELETE_FAILED;
+                $this->backup->save();
+
+                SocketEvent::dispatch(new SocketEventDTO(
+                    projectId: $projectId,
+                    type: 'backup.updated',
+                    data: new BackupResource($this->backup),
+                ));
+
+                return;
             }
 
             $this->backup->delete();
@@ -45,5 +59,19 @@ class DeleteJob implements ShouldQueue
     public function failed(Exception $e): void
     {
         ServerLog::log($this->backup->server, 'delete-backup-failed', $e->getMessage());
+
+        if ($this->backup->exists) {
+            $this->backup->status = BackupStatus::DELETE_FAILED;
+            $this->backup->save();
+            $this->backup->files()
+                ->where('status', BackupFileStatus::DELETING)
+                ->update(['status' => BackupFileStatus::DELETE_FAILED]);
+
+            SocketEvent::dispatch(new SocketEventDTO(
+                projectId: $this->backup->server->project_id,
+                type: 'backup.updated',
+                data: new BackupResource($this->backup),
+            ));
+        }
     }
 }
